Cleanup of the cli/cmd pkg
